build: add build scripts
