+++ .gitignore
@@ -6,3 +6,5 @@ node_modules
 .yarn
 .pnp.cjs
 .pnp.loader.mjs
+# project build
+build
