--- index.ts
@@ -1,4 +1,0 @@
-import * as dotenv from "dotenv";
-// load environment variables (if you have them)
-dotenv.config();
-console.log("Welcome to ts-template");
