--- scripts/esbuild.ts
@@ -1,3 +1,0 @@
-import esbuild from "esbuild";
-import { esBuildContext } from "./esbuild-config";
-esbuild.build(esBuildContext);
