--- scripts/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "useDefineForClassFields": true,
-    "allowJs": false,
-    "esModuleInterop": true,
-    "module": "CommonJS",
-    "allowSyntheticDefaultImports": true,
-    "strict": false,
-    "forceConsistentCasingInFileNames": true,
-    "target": "esnext" /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */,
-    "lib": ["esnext", "dom", "dom.iterable"] /* Specify a set of bundled library declaration files that describe the target runtime environment. */
-  },
-  "include": ["./helpers", "./constants", "./index.ts"],
-  "exclude": ["./esbuild.ts", "./esbuild-config.ts"]
-}
